no artefacts

+++ gridReader/gridReader/Calculator.cpp
@@ -104,7 +104,7 @@ double Calculator::calcTetraVolume(double a[3], double b[3], double c[3], double
 double Calculator::calcAverageEdgeLength(vtkIdType tetra, vtkUnstructuredGrid *tetraGrid) {
     double sum = 0;
     for (vtkIdType edge = 0; edge < tetraGrid->GetCell(tetra)->GetNumberOfEdges(); edge++) {
-        sum += calcEdgeLength(tetraGrid->GetCell(tetra)->GetEdge(edge)->GetPoints()->GetPoint(0), tetraGrid->GetCell(tetra)->GetEdge(edge)->GetPoints()->GetPoint(1));
+        sum += calcEdgeLength(tetraGrid->GetCell(tetra)->GetEdge((int)edge)->GetPoints()->GetPoint(0), tetraGrid->GetCell(tetra)->GetEdge((int)edge)->GetPoints()->GetPoint(1));
     }
     return (sum / tetraGrid->GetCell(tetra)->GetNumberOfEdges());
 }
